# Conflicts:
#	src/mod/data/areas.h
#	src/mod/data/features.h
#	src/mod/data/zones.h
#	src/mod/externals/Dpr/UI/BoxStatusPanel.h
#	src/mod/externals/Dpr/UI/BoxWindow.h
#	src/mod/externals/Dpr/UI/ContextMenuItem.h
#	src/mod/externals/Dpr/UI/ContextMenuWindow.h
#	src/mod/externals/Dpr/UI/UIManager.h
#	src/mod/externals/FieldManager.h
#	src/mod/externals/PlayerWork.h
#	src/mod/externals/Pml/PokeParty.h
#	src/mod/externals/System/Collections/Generic/List.h
#	src/mod/externals/System/Func.h
#	src/mod/externals/UnityEngine/Events/UnityAction.h
#	src/mod/features/commands.cpp
#	src/mod/features/commands/commands.h
#	src/mod/features/main.cpp
#	src/mod/ui/components/element_inspector.h

+++ src/mod/romdata/zone_rates.cpp
@@ -24,7 +24,7 @@ int32_t Roll(nn::vector<uint32_t>* rates)
     int32_t currentSum = 0;
     for (size_t i=0; i<(*rates).size(); i++)
     {
-        if (roll < currentSum + (*rates)[i])
+        if (roll < currentSum + (int32_t)(*rates)[i])
             return i;
 
         currentSum += (*rates)[i];
